impulse_autopilot/requirements.txt
# FastAPI Backend
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Task Scheduling
apscheduler==3.10.4

# AI Integration
openai==1.12.0
httpx==0.27.0

# MT5 Communication
requests==2.31.0

# Data Processing
pandas==2.2.1
numpy==1.26.4
pyarrow==15.0.0
polars==0.20.10
duckdb==0.10.0

# Visualization
plotly==5.19.0
matplotlib==3.8.3
seaborn==0.13.2
mplfinance==0.12.10b0
altair==5.2.0

# Data Analysis & Technical Indicators
pandas-ta==0.3.14b0
statsmodels==0.14.1
scikit-learn==1.4.0
scipy==1.12.0
xgboost==2.0.3
QuantStats==0.0.62

# WebSocket
websockets==12.0

# File Upload
python-multipart==0.0.9

# HuggingFace Sync
huggingface_hub==0.20.3

# Yahoo Finance
yfinance==0.2.36

# Crypto (CCXT)
ccxt==4.2.8

# Web Search
duckduckgo-search==5.1.0

# Utilities
cryptography==42.0.5
toml==0.10.2
